+++ .gitignore
@@ -19,5 +19,6 @@ mps/org.modelix.model.client/lib
 # modules are configured by gradle
 .idea/modules
 modules.xml
-modelix.version
 .DS_Store
+
+/modelix.version
