+++ model-api/build.gradle
@@ -39,7 +39,6 @@ kotlin {
     sourceSets {
         commonMain {
             dependencies {
-                api project(':model-api')
                 implementation kotlin('stdlib-common')
             }
         }
@@ -122,21 +121,4 @@ spotless {
                 ' */\n' +
                 '\n'
     }
-}
-
-task copyModelClientToMps(type: Sync) {
-    from configurations.jvmDefault
-    from "$buildDir/libs"
-    into "$projectDir/../mps/org.modelix.model.client/lib"
-    rename { fileName ->
-        stripVersion(fileName)
-    }
-    exclude 'log4j*.jar'
-    exclude 'annotations*.jar'
-}
-
-assemble.finalizedBy(copyModelClientToMps)
-
-String stripVersion(String fileNameWithVersion) {
-    return fileNameWithVersion.replaceAll(/-\d+(\.\d+)*(-SNAPSHOT)?/, "")
 }
