+++ model-api/src/commonMain/kotlin/org/modelix/model/PlatformSpecific.kt
@@ -1,0 +1,9 @@
+package org.modelix.model
+
+import kotlin.reflect.KClass
+
+expect fun logError(message: String, exception: Exception, contextClass: KClass<*>)
+expect fun logDebug(message: () -> String?, contextClass: KClass<*>)
+expect fun logTrace(message: () -> String?, contextClass: KClass<*>)
+expect fun logWarning(message: String, exception: Exception, contextClass: KClass<*>)
+expect inline fun <R> runSynchronized(lock: Any, block: () -> R): R
