+++ model-api/src/commonMain/kotlin/org/modelix/model/api/CompositeNodeResolveContext.kt
