+++ model-api/src/commonMain/kotlin/org/modelix/model/api/IConcept.kt
