+++ model-api/src/commonMain/kotlin/org/modelix/model/api/IReadTransaction.kt
