+++ model-api/src/commonMain/kotlin/org/modelix/model/api/ReadTransaction.kt
