+++ model-api/src/commonMain/kotlin/org/modelix/model/area/CompositeArea.kt
