+++ model-api/src/commonMain/kotlin/org/modelix/model/area/IAreaListener.kt
