+++ model-api/src/commonMain/kotlin/org/modelix/model/util/pmap/COWArrays.kt
