+++ model-client/src/jsMain/kotlin/org/modelix/model/PlatformSpecific.kt
@@ -6,23 +6,6 @@ actual fun sleep(milliseconds: Long) {
     TODO("Not yet implemented")
 }
 
-actual fun logError(message: String, exception: Exception, contextClass: KClass<*>) {
-    console.error(message, exception)
-}
-
-actual fun logWarning(message: String, exception: Exception, contextClass: KClass<*>) {
-    console.warn(message, exception)
-}
-
-actual fun logDebug(message: () -> String?, contextClass: KClass<*>) {
-    val msg = message()
-    if (!msg.isNullOrEmpty()) console.log(msg)
-}
-
-actual fun logTrace(message: () -> String?, contextClass: KClass<*>) {
-//    console.log(message())
-}
-
 actual fun bitCount(bits: Int): Int {
     var i = bits
     i -= (i ushr 1 and 0x55555555)
@@ -45,8 +28,4 @@ external object uuid {
 
 actual fun randomUUID(): String {
     return uuid.v4()
-}
-
-actual inline fun <R> runSynchronized(lock: Any, block: () -> R): R {
-    return block()
 }
