+++ settings.gradle
@@ -2,6 +2,7 @@ rootProject.name = 'modelix'
 
 include 'ui-client'
 include 'model-server'
+include 'model-api'
 include 'model-client'
 include 'mps'
 include 'ui-server'
